moved pages to separate folder
*moved pages to pages/ folder
